Stop overflowing the depth calculation in tt::should_overwrite
Caught while running some games in debug mode. Shouldn't make a
difference in practice as it'll only happen when we hit depths near our
max, so I assume this is a fluke:

Elo   | 4.92 +- 4.30 (95%)
SPRT  | 4.0+0.04s Threads=1 Hash=16MB
LLR   | 2.98 (-2.25, 2.89) [-5.00, 0.00]
Games | N: 5582 W: 1372 L: 1293 D: 2917
Penta | [30, 468, 1711, 557, 25]

+++ engine/src/engine/transposition_table.rs
@@ -271,7 +271,7 @@ impl TranspositionTable {
 
         // Weight entries by depth searched, preferring newer entries unless old entries are searched
         // to significantly higher depths
-        if new.depth + 4 > old.depth {
+        if i32::from(new.depth) + 4 > i32::from(old.depth) {
             return true;
         }
